requirements.txt
torch
clip-anytorch
numpy
opencv-python
faiss-cpu
streamlit
Pillow
# torch.classes.__path__ = []